New OffsetWriter and *List classes
Some parts of the map, especially the RGN part, are written
out very early in the process. In the road's case, the
offsets of the RoadDefs aren't known yet, because those
offsets depend on later (not yet written, not yet "thought
of") entries in the RGN section.

The current solution is to keep a list of positions in the
RGN, that need rewriting with the corrected offset. That's,
what OffsetWriter does. The only add-on it also has: OR the
offset with a mask. It's common to have OR'd pointers, so
OffsetWriter supports it.

OffsetWriterList is even simpler: It's just a list and does
the final "call all entries in the list with one offset to
be written".

It's not clear, if this is the best approach to the
problem, but it works for me currently.