Let MapElements have a generic "UserData" attribute
MapLines are the generic representation of roads. Before
they become Polylines (the ones in the RGN). On the way
there, MapLines get transformed somewhat. This can include
smoothing of the coords for the current resolution, or
splitting into parts for different Subdivisions.

One road should have one RoadDef for it. No matter, in
which Subdivision, resolution, or what it is written
finally. All of them should have _one_ RoadDef.

So the _one_ RoadDef is linked from the top MapLine.
When new MapLines (for whatever reason) are created, they
will inherit this link. So that the final MapLine, which
gets then converted to a Polyline (for the RGN) still knows
about the RoadDef.