Add a noname style.
Can be used with --style=noname

Major roads that do not have a ref, as well as residential roads that do
not have a name are highlighted.  You can use the tag noname=yes to
show that a road doesn't have and name and therefore shouldn't be highlighted.

No-name roads show up as a thick line that is visible at higher zooms than is
usual for residential roads to help you find areas with poor naming coverage.