Allow for extremely uneven distribution of nodes in subdivision.
The last commit croaked on subdividing a region if all of the region's
nodes went into the same sub-region. That is overly restrictive because
you could possibly have a rectangular region where the nodes are only in
one half of the region.

This new commit now detects when a sub-region becomes impossibly small
rather than worrying about the distribution of nodes in the sub-regions.