Add highway shields and other ways of modifying values in the style files.
Implemented as a series of modifiers that can be used in the style files.

HighwaySymbolFilter: inserts the highway symbol markers after cleaning up the name.
HeightFilter: Prepends the value that makes the garmin display as an elevation.
SubstitutionFilter: allows arbitary substitutions within a value.
PrependFilter: Prepend one the garmin magic values to a string.

Patch by Toby Speight as the end result of a discussion involving a number of people see
http://www.mkgmap.org.uk/pipermail/mkgmap-dev/2009q2/001573.html for the
full details.

This only appears to work with 8bit outputs.