Add ability to make GType directly from tags without any styling.
If mkgmap:gtype tag is present, its value is expected to be a comma
separated list of the form: kind,code,minres,maxres,roadclass,roadspeed.
Where kind is 1 for points, 2 for lines and 3 for areas. The first 3 values
are required, the others are optional.

Also, decodes ~[0x??] constructs in name strings to produce highway
shields, etc.