Added support for extended types.
These are sometimes called '3-byte types' but the upper 8 bits of the
type is not encoded in the map objects (just the lower 16 bits) so any
non-zero value for the upper byte has the same effect.

Further work is needed to encode the extra bytes that specify stuff like
depth and colour + all the detail associated with marine lights.