Start a branch for the global index work (mdr).
The plan is to quickly create the minimum amount that works, merge it back
and then repeat.  I am going to read img files to create the index as there is
lot of information that is needed in them and it would be useful to create
one separately anyway.  A similar decision was made wrt the overview map,
and so the same code can be used to finally implment a better overview map too.