Reduce number of Table A entries.
Previously, Table A entries were unique for each combination of start node,
end node and RoadDef. But no information from the start and end nodes goes
into Table A so we can ignore the start and end nodes and all roads
within a given routing centre that use the same RoadDef can now share a
Table A entry. This reduces the number of entries required which makes
the .img files smaller and, hopefully, also improves the routing because
routing centres can now cover a larger area.