Add a character set aware label decoder.
This is a fairly big change as the CharacterDecoder interface is changed to have
a method that returns a DecodedText rather than an EncodedText.  The new class
DecodedText is more useful as it contains the actual string that you need.

non-ascii characters will now be correctly represented in the index, although they will still not
work as the sorting needs adjusting.