Add ability to set a road's class and speed from a CoordPOI.
If the way has a point with a POI that has mkgmap:road-class or
mkgmap:road-speed tags they will be transferred to the way along with any
mkgmap:road-(class|speed)-(min|max) tags. The tag values will be used to
modify the way's class/speed parameters. Furthermore, the way will be
split at the points before and after the POI to limit the region that has
the class/speed modified.

The mkgmap:road-class and mkgmap:road-speed tag values can either be an
absolute integer value or they can be prefixed with + or - to indicate
that they are relative to the values set by the style rules. The min/max
tag values are all absolute values.