If a polish map contains any background polygons, don't generate one.
Normally, a polygon of type 0x4b is automatically generated to cover the
map's background. For OSM maps that can't explicitly contain such a
polygon, this is reasonable. Polish maps can contain polygons of type 0x4b.
So if a Polish map does contain any polygons of type 0x4b, suppress the
automatic generation of the background polygon.