Avoid assertion when splitting way at a previous node.
When a way is split to avoid exceeding the maximum number of points in
a way and the split is done at a previous node, set a flag so that the
later code that checks for duplicate node indices doesn't barf because
the node index has already been seen before. In fact, in this situation,
we can reverse the check such that it requires the node index to have
been seen before and it will barf if it hasn't.