Fix regression introduced by r1435.
When scanning backwards looking for a place to split the way to limit
the way's number of points, don't set the flag that says we're splitting
the way at a previous node if the split is going to be at the current
point (i.e. we didn't need to go backwards at all).