Bug fixes - tag removal crash and excessive table expansion.
No longer decrements size when a tag is removed as this causes a
problem because reducing size makes it look as if there is more
space available than there really is as the removed tag's key
is still taking up a slot and will do so until the arrays are resized
by the next call to ensureSpace().

Also, ensureSpace() now zeros size before copying the existing tags/values
to the new arrays so size will be correct when the method returns.