Can now close small gaps in the coastline.
Added generate-sea sub-option, close-gaps, to specify the maximum length
(in metres) of gap that can be closed. Default behaviour is not to close
any gaps.

Gaps between coastline segments are closed with a straight line.

A separate line tagged natural=mkgmap:coastline-gap is also generated
so that you can see where the gaps have been closed.