Fix problem where polygon rules would overwrite line rules with the same key.
Make the style tester on osm files more useful by using preserve-element-order so files are directly
comparable and preserve the order of the elements in the input files.

Default style now appears to fully work in the style branch.
It does not use continue, and that is not compatible yet.