Add new functionality to the turn heading adjustment code.
It used to do one adjustment, which was to ensure that the side road
heading change relative to the heading of the outgoing main road segment
was greater than some minimum amount.

It now also ensures that the side road heading change relative to the
incoming main road segment is also greater than some minimum amount.

Furthermore, an optional bit mask can be provided to choose which of these
adjustments are enabled:

--adjust-turn-headings[=BITMASK]

	Optional BITMASK (default value 3) allows you to specify which
	adjustments are to be made (where necessary):

	1 = increase angle between side road and outgoing main road
	2 = increase angle between side road and incoming main road