I've taken a different approach and no longer collect rules into different buckets to form an index.
The rules are now read literally from the top and applied to the element in turn.  It turns out
that this does not add any significant time because it is still very quick compared to
everything else that is going on.

Continue is not enabled at all in this revision so do not use if you need it.