More specific warnings for multipolygons (intersections, exchanged inner/outer roles).
Self intersecting polygons are now handled as multiple polygons instead
of using only the first polygon

Inner polygons that are not contained by any outer polygons are
dismissed now. There is a high probability that inner/outer roles are
exchanged or that the inner polygon does not belong to the multipolygon.

Remove polygon tags from unclosed ways because sometimes they are closed
later in mkgmap.

WanMil