Merge the style branch.
The style language will now behave as documented in terms of
the order in which rules and actions are run.

An action will be run after all matching actions above it in the file
and before all matching actions below it.  Previously they were run in
an indeterminate order.

The effect of an action will also only be felt by rules below it.
Again, this was previously not true.