Drop all tags from the osm file that are not used in the applied style.
Whether this makes a big difference or not depends on the country.
In the UK which is mostly manually taged it doesn't make a great deal
of difference.  There are only an average of 2 tags per way and about zero per
node.

In countries with data imports, there are typically many tags per way (and even nodes)
that refer back to the original data.  Having tags on nodes is particularly bad for memory
consumption since there are more nodes than ways and most are not POIs.  We are forced
to allocate a Tags object even though there is likely to be nothing to do as none of the tags
are used.

Denmark is the worst in this respect and so will show the best improvement with this patch.
See also: http://www.mkgmap.org.uk/pipermail/mkgmap-dev/2009q3/003597.html

For this to be successful you need an acurate list of all the tags that could be used.  Tags are
examined not just in the style itself, but there are hardwired tag lookups in the style system itself
and there are non-style related usages as well.  For the moment all the built in usages of tags
are held in the builtin-tag-list file.  I think I have got them all for normal use, but there may be
options that require ones I've missed.