Merge the style-speed branch to trunk.
This drops all tags that are not used right at the point of reading from the
input file.

This has varying benefits depending on the input data.

In the best case, we avoid creating Node objects for a large number of points
that should just stay as Coord objects.  As the Node object contains numerous
sub objects such as the Tags object this saving can be very large.

Even on ways, where it is rarely possible to avoid creating a Tags object, you can reduce the number of resize operations needed.

The biggest benefit is for Denmark and countries that have had automated bulk
imports.  In the UK the difference is merely in the 10s of megabytes rather than
the hundreds or thousands.