Use road read from net and not from reading in rgn.
Sort regions according to the Sort in the img files. (recompile tiles needed).

Sort cities with a secondary sort on region in the img files. This fixes an uncommon problem
with losing a city from the drop-downs. Not sure if I am just chasing the problem around, but this
seems to work in lots of cases where it didn't before.  I have only seen the problem on
one city though - so hard to tell.