Setting the sort parameters from the sort description file.
Now everything is set from the cp1252.txt file and there are no hard
wired parameters.  I don't know if you *can* change them and still have a
working map, but perhaps someone would like to try :)

It is well established if the id numbers are different in the index,srt and img files
then things will not work.