Temporarily disable mdr8 for the "Calle problem".
MDR8 is an index of the first four letters of the street name to the record
number every 10,000 records or so. This causes a problem when the word for
street comes first in the name and there are different ways of spelling that word.

So in Spain many streets begin with Calle, so many that there are several repeated
index slots for "CALL".  I think that in itself would not be a problem (or at least it would
probably be easy to fix), but in some cases it is spelt as "Çalle" (C with a cedilla) and this
confuses the algorithm to find the first match.

It may be easy to fix, but will need more investigation.