Reduce artefacts caused by multipolygon cuts
The artefacts are avoided by selecting cutting lines that lie on a n*2048 Garmin units grid. 
Additionally a bug is fixed where in some special cases an inner area was not cut out. This was caused by the java.awt.geom.Area.equals() method, that sometimes returns true although the compared areas are not congruent.