Improved street search and city search.
There is an entry in the city table for every separate city based on city/region/country
names. If the same city/region/country occurs in more than one tile, then there is an
entry in each tile.

There is a new mdr20 section for each city/region/country ie. there is not a new one for each
tile that the city/region/country occurs in.

Cities also occur in the POI index and the same rules apply there.

Merge from city-region-index branch so see that for more info.