Apply fix being tested for trunk for sorting of mdr20.
That fix massively increases memory use, so also deal with that here.
 
Generalise the caching system previously used on mdr21 and 22 so that it can be used here
too. Since the new cache is used inside of createSortKey() we don't need the extra
MultiSortKey() objects in mdr21 and 22 either, so they now need less memory.

Since this change includes a bug fix, the generated index file is different.