Avoid NullPointerException in BoundaryUtil when loading a preprocessed bounds file.
The Exception occurred when loading a bounds file that contains a boundary which first area is counterclockwise due to a problem with the Java2DConverter which has to round points to integer coordinates. In rare cases this changes a polygon from clockwise to counterclockwise.
Now issue a warning if this happens and ignore the polygon.