The osmpbf dependency was erroneously marked with changing="true".
This tells Ivy that the jar might change
despite the version number not changing, so always check the
repository.  This is usually used to express a dependency on the
latest snapshot (e.g., nightly build) of a 3rd party project.  This is
not the case with osmpbf; its version number represents a fixed
revision, so 'changing' should be false.

-Richard Hansen