Add precomp-sea option
It is now possible to generate the sea areas using precompiled tiles (OSM or PBF format). The tiles use a raster size of 32768. A mandatory index file gives the file names for the tiles. Tiles with land or sea only are not stored but are marked in the index with "land" or "sea".  The precomp-sea option gives the directory with the precompiled tiles and the index file.