Reduce maximum memory consumption
At several places lists and other variables are cleared and set to null at an earlier stage so that they can be GCed. The main change is to allow the ElementSaver to drop the coordMap before the OSM hooks start. Some tests have shown that the maximum memory requirements decreases by 9%. But maybe this is not always the case.