Always copy the point list when creating a new Way object
This prevents unintended use of the Way constructor which didn't copy the points list. If the points of a way was changed afterwards it happened that the points list of a copy was changed also. This caused routing errors in special situations (e.g. a road was also a member of a multipolygon that has a highway tag too).