Enable coastline drawing when using precompiled sea
The coastline ways are not contained in the precompiled sea. But they can be taken from the OSM input tiles. So when the --precomp-sea option is set all costline ways are copied and additionally marked to be used only with the lines style file. This let the coastline reappear with precomp-sea.