House numbers, edge cases
Including zero. However a zero does not show up on the map. But it otherwise
works - number 2 is a little further down the road as you would expect - it just
doesn't display the number zero. In all or most parts of the world zero is not
a valid house number, so this is more about preventing a crash on invalid data.