Fix 'include' statement that is the very first statement in an included file.
The very first statement in an included file was not being tested to see if it
was an include command. As a result the 'include' token was was seen as
a tag name and would lead to a syntax error.

Now restart the conditional processing when ever a new file is included.