Do not always set drive-on-left flag to false if option precomp-sea is set
In case a tile has mixed sea/land content and uses the precomp-sea option the precompiled sea tiles are loaded by an OsmMapDataSources instance. This creates a an option less StyledConverter which sets the drive-on-left flag to false. This overrides the drive-on-left/right option of the config file. 
This change adds an internal option "ignore-drive-on-left" which does not initialize the drive-on-left flag in the StyleConverter. This option is only set by the SeaGenerator when loading precompiled sea tiles.