Changes to code-page and TYP .txt files.
If the .txt file begins with a BOM character, then the file is in
utf-8 (or -16, but we don't support that) so do not try to read it
in the CodePage character set and encoding.

Also the --code-page option completely overrides any CodePage directive in
the file.

The advice now is that .txt files should be in utf-8 and there is no need for
a CodePage directive.