Use the actual transliterated characters to sort on.
We now save the characters that are placed into the file in their codepage
representation. (but not in the 6-bit format).

Still only 8-bit codepages supported, so all the top bits of the characters
will be 0.

This is a transition. New and old ways of doing things co-exist and the usage
of the old way will be gradually removed, while keeping everything working.

There may be debug statements that will cause an exception if turned on.