Merge from the sort branch.
This solves multiple problems searching for addresses when the default
ascii character set is used, or when many characters cannot be 
represented in the target character set and have to be
transliterated.

Complete sort tables for code-pages: 1250, 1251, 1252, 1253, 1254,
1255, 1256, 1257 and 1258.

A few corrections to the sorting algorithm so that sorting by
sort-keys always gives the same result as sorting using the Collator.

If anyone uses there own sort descriptions, then there is one change:
everything is in terms of unicode. The file itself should be in utf-8
as before, but all characters with numeric references should be in
unicode, rather than the target character set.

There are numerous internal changes that are the beginnings of
unicode support, but these changes are not visible, so unicode
maps will still not be searchable.