Size of 25 ptr field is determined by size of section 5.
Sect 25 is cities sorted in order of country. If every city has a country
then the sizes of 5 and 25 will be the same.  If we have cities without a country
then sect 25 can be smaller.

It appears that the mdr25 pointer in mdr29 is sized based on the total number of cities
and not the number of cities that have a country assigned.  Of course with perfect
data these would be the same, but we don't always have that.