improve WrongAngleFixer: before merging nodes that are not very close, check that a nearly straight line goes through both nodes and that this line is improved by the merge.
In rare cases, nodes where merged causing distorted lines.
The merge should happen in a situation like this:
          b  
          b
aaaaaaaaaaaaaaaaaaaaaaaaaaaa
            c
            c

Road a is straight, roads b and c are connected to it with
nodes that are close together. If a is not perfectly horizontal,
the two connection points might be rounded to very different 
positions, so that road a is zig-zagging in the map.
The merge helps to avoid this.