improve LineMergeFilter: Don't merge lines for higher levels
LineMerger is called with a list that contains all lines, also 
those which are not written at current resolution. The old 
versions merged all of them where possible, but only the
ones with the right resolution were written to the img file.
The new version skips the merge for all lines which are not
written at the current resolution.

This has different effects:
1) better performance (a few % faster), because the work is not done
multiple times, and numbers of lines to merge are much smaller on 
lower resolutions (e.g. 25 instead of 250.000 at resolution 14)
2) the merging of lines for resolution > x preserved some points 
for resolution x which are now not preserved. This has a significant
impact on output, many shapes and some lines look different. Hard to say 
if the new result is better.
3) the img size is a bit smaller (maybe only a few bytes)