Style filter arguments can now be quoted.
For backward compatibility you don't have to quote arguments that contain
spaces or symbols but its a good idea to do so from now on. The documentation
has been updated to follow this guideline.

You can now use a pipe symbol in a regular expression which there
was no way of doing before.

Eg:  ${name|subst:"(Dr|Doctor|Dokter) ~>Doktor "}