work around for edge case reported by Steph
After adding a new point to a road the distance to the road
can be a few cm higher. If the original distance is very close
to the maximum (150m), the new distance might be too high.
In this case we should use the max value and continue to use the 
house.