remove wrong rules to set mkgmap:city for France
This is how it should work : in France, we don't use admin_level=9 expect for 3 cities (Paris, Lyon, Marseille).
For an address in these 3 cities, the city name is still the one from admin_level=8 but the post code is from admin_level=9. We don't use admin_level=9 boundaries for anything else.
- Paco Tyson