sharp-angles-v4.patch with small modifications
- remove the code for --adjust-turn-headings which no longer worked
- add code to fix sharp angles (enabled by default)
Enlarge sharp angles at road junctions so that the time penalty
is smaller. In OSM, junctions are often drawn without reflecting
the exact angles of the roads, they are more symbolic. We assume 
that in most cases an angle below 45° doesn't really exist as such
and therefor the heading values of the arcs are modified so that the 
Garmin routing algorithm "sees" a less sharp angle.
If needed, the option --ignore-sharp-angles can be used to disable this.