check_roundabout_connections_v4.patch : More checks for the check-roundabout option
Mike Baggaley:
Hi Gerd, please find attached a version 3 patch, that seems to me to provide everything we need at the moment. 
It provides two warning messages  of forms “Roundabout is connected to more than one road at…” and “Roundabout is connected to 
a road and other highways at…”. 
It does not warn for multiple highways where there is no road with car access 
(e.g. a cycleway that goes straight across a roundabout).