Update the MPS file for unicode names.
This is the least clear part of updating to unicode. As the gmapsupp has no codepage field
and it is possible (at least in theory) that different families within the gmapsupp have
different code pages, it is not clear what to do.  Mapsource/BaseCamp appear to keep
the same character encoding as in the tdb/img, but I've not tried with different codepages.