Implement a new map output format "gmapi"
When the --gmapi option is given then the map will be created
in a directory named after the family name with an extension .gmap
This format is used to install maps on a Mac.

As with the --gmapsupp option, you can start with a set of .img files that
where created in a previous mkgmap run if you wish, as well as in one
step from the source OSM .pbf, .osm.gz etc files.