echo.patch by Mike Baggaley
This change improves the format of the messages such that the type of element is
included and the fake Id is replaced with a fragment saying the element has
been generated. The latter allows the output of different runs of mkgmap to
be compared without randomly generated Ids causing differences. I can't see
any need to know the actual fake Ids, just the fact that the element has
been generated should be sufficient. I have also removed "- " from echotags
and ":" from echo as unnecessary and inconsistent.

Existing message:
4611686018654788095 (1326812) -
[amenity=school,mkgmap:area2poi=true,name=School] Description as name

New message:
MultiPolygonPOINode generated from 1326812
[amenity=school,mkgmap:area2poi=true,name=School] Description as name