fix rare severe error message "internal error, distance to road too large,..."
Happens when rather long road segments at certain bearings should be split for house number search but no Garmin point is close enough to prevent zig-zagging.

The current algo calculates a wanted split location and tries to find a good Garmin
point close to it. It might be better to check if any good Garmin point is an improvement.