ShapeSplitFix by Ticker Berkin
The patch is just to fix a copy/paste error.

This code deals with a pathological case where there is a cut through a
shape that loops around and along itself (3 lines on top of each other)
and to distinguish whether it is a hole looping back inside or an area
looping back outside.