Allow the FileNode access to some of the ImgFileWriter details.
To be able to calculate the block size and the size of the directory
we need the file sizes before they are written out.  Then we need
to be able to control the sync of the buffer from the FileNode.close()

Preparation only, there should be no actual change in this commit.