RulesTest: clarify that all errors found are bugs in mkgmap
In previous versions it was not possible to tell if an expression
should work or not, so some syntax errors reported by this test
program were not really bugs.

In the current version only expressions that should
work are generated so any error or crash found is a mkgmap bug.

There is now no need to treat SyntaxException differently as they
are always bugs as well as crashes.