revert most changes from r4223 in WrongAngleFixer
This is a work around for a possible Assertion Error which can happen
when a node on a country boundary is very close to the tile boundary. The soft clipping process creates an additional node exactly on the tile boundary. This might cause a zig-zag.
The node on the tile boundary must not be replaced by the other one because the real clipping would add it again.