Add check to make sure that the style is compatible with the levels option.
A style should not add objects with a minresolution of 24 when the levels options defines a lower resolution for level 0.
Unfortunately this only works well when the option is specifiied in the options file of the style. In this case all problematic lines are listed.
If the user adds an option like --levels=0:23,1:22,2:20,3:18 it is ignored when --check-styles is used and only one error message is produced, eg.
Error in style: Error: Cannot use type [0x2f12 resolution 24 continue] with level 0 at resolution 23 in style file points, line 48
Could not open style