Fix error: Shapes are distorted when 2nd last and last point are at same position or when 1st and 2nd point are at same position (1st and last point are identical)
This is rather unlikely with OSM data but (since r4360) it happened always with polish input as it removed code
which caused coords at the same place to be the same object. I removed that in r4360 because I think it can cause problems with housenumber nodes or routing nodes.
So, we have an old error which didn't show up until the change in r4360. 
This change fixes the old error and changes the code in the polish reader so that it doesn't produce this special case.