improve evaluation of barrier=* nodes (barrierAccess.patch by Ticker Berkin)
Default style "points" has code that, when option --link-pois-to-ways
is in effect, sets highway access and speed restrictions that mkgmap
then imposes on the highway that the point is on.

These rules had various errors:
- didn't expand the access mode hierarchy, see
  https://wiki.openstreetmap.org/wiki/Key:access
- tested the wrong tag for mkgmap:bus and mkgmap:delivery
- if the "access" tag was set, apply this to all modes before
  processing allowances and restrictions imposed by the
  barrier; but these actions would have no effect because the
  the specified mode tags are now all set
- assumed that a gate stops motor_vehicle access
- didn't handle other common forms of gate
- allowed bicycle through kissing_gate and stile

This patch fixes these problems and also adds a speed restriction for
all barriers.

The general principle is that barrier rules add specific "no"
restrictions that the barrier prevents, but must not add "yes".
The "access" tag is a default for all modes not implicitly or explicit
handled; there could be, say, osm tags [barrier=stile,
access=destination] and this should not imply there is unrestricted
foot access.